Autodock Vina automation using BASH
You can use this script for automating Autodock Vina during a selected number of iterations. The ligand files must be numbered in sequential order, e.g: L1.pdbqt, L2.pdbqt, L3.pdbqt, L4.pdbqt, and so on. The receptor.pdbqt used must be placed in the same folder as the ligands.pdbqt files.